Introduce Noload Gzip Compression Acceleration
Noload compression provides a way to offload GZip compression to an
FPGA accelerator(s). The user can enable GZip compresion to be offloaded
to Noload accelerators by setting the zfs module parameter
zfs_noload_compress_enabled to 1. By default this module parameter is
disabled. This allows for QAT compression offload to still happen if
Noload GZip compression is not enabled.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>
Authored-by: Logan Gunthorpe <logan.Gunthorpe@eideticom.com>