Introduce NoLoad Gzip Compression Acceleration
NoLoad compression provides a way to offload GZip compression to an
FPGA accelerator(s). The user can enable GZip compression to be
offloaded to NoLoad accelerators by setting the zfs module paramater
zfs_noload_copmress_enabled to 1. By default this modules paramter is
disabled. This allows for QAT compression offload to still happen if
NoLoad GZip compression is not enabled.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>
Authored-by: Logan Gunthorpe <logan.Gunthorpe@eideticom.com>