Adding Falling back to SW Compression for NoLoad
In the event there is a failure or no NoLoad compression cores we need
to fall back to SW compression, but still mark the BP that NoLoad
compression was used. This way, we can make sure that if we only
have decmpression cores we can still use the NoLoads to decompress
data when the dataset property is set to gzip-noload.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>