Fixing dbuf dirty records in O_DIRECT writes
We need to make sure that we are removing any dirty records that
might be associated with the same ARC buf that we could possibly destroy
in dmu_write_direct_done(). In order to do this, we also have to make
sure the current dirty record is no being sync'ed out with a shared ARC
buf. These changes address that by iterating through the dirty records
of the dbuf and make sure we can either directly remove the shared ARC
buf in dmu_write_direct_done() or if the ARC buf has already been
destroyed for us if the data has has been sync'ed.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>