Updating a few issues with O_DIRECT
Updating semantics of which block pointer to use in dbuf_read():
In the event that a Direct IO write has occurred the block pointer for
the write is in the dirty record. In order to make sure the proper block
pointer is used in dbuf_read(), the dirty record is checked to see if
contains an updated block pointer. We want to return the most recent
version of the block pointer, so we make sure if a Direct IO write has
updated the data on disk the read will use that block pointer to grab
the data. However, if the dirty record txg is being synced, we need to
wait to make sure the dbuf's state does not become inconsistent between
an open context read and a Direct IO write syncing it's metadata to
disk.
This is needed because in dbuf_read_impl() we change the dbuf's state to
DB_READ and then drop the db_mtx. This can lead to the dirty record
being sync'ed out with a dbuf state of DB_READ in dmu_sync_leaf(). This
is an invalid state in sync'ing context.

Added kstat's for dbuf Direct IO modifications:
Added kstats for
 - direct_undirty_wait
 - direct_undirty
 - direct_read_wait
 - dirct_read_retry
So we can do accounting for the corner cases where buffered and direct
operations are mixed.

Updating docs:
Updated zfsprops man page for the direct property and the dbuf state
transition diagram in dbuf.h for the dbuf states to include Direct IO.

Style changes:
Cleaned up some documents and parens.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>
Co-authored-by: Brian Behlendorf <behlendorf@llnl.gov>