Updating test structure for O_DIRECT
I moved the O_DIRECT tests to their own test directory and cut down
the size of each test to make sure the tests run for a shorter amount
of time. I also added a tag (direct) for the zfs-tests so the tests can
now be started by doing:

./scripts/zfs-tests.sh -T direct

I also updated the dbuf Direct IO kstats placement so the kstats are in
alphabetical order.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>