Fixing some errors
Removed the extra call of zfs_uio_free_dio_pages() in zfs_read() to stop
the PTE faults that were occurring on O_DIRECT reads.

I also updated how we were accounting for O_DIRECT reads and writes. We
were double counting for O_DIRECT writes in
dmu_buf_hold_array_by_dnode() when doing reads. Because this function is
called for both O_DIRECT read and writes we have to make sure not to
account for writes when doing reads.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>