Had to update how zfs_setup_direct() is used
Until the issue with EFAULT being returned from FreeBSD is resolved I
had to change how we use zfs_setup_direct(). It no longer returns
EAGAIN. This means that Linux has to check the flags for O_DIRECT after
calling zfs_setup_direct() to determine if the it should still do
O_DIRECT without an error returning from zfs_setup_direct().

Using this update, all the direct tests pass in FreeBSD. This is mostly
just to mask the issue with EFAULT at the moment and this issue really
must be fixed on the FreeBSD side.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>