Reverted zfs_setup_direct() to previous version
Finally found the mistake I made with zfs_setup_direct(). In the case
of os->os_direct == ZFS _DIRECT_DISABLED I was not calling
ZFS_EXIT(zfsvfs) before returning EAGAIN. This of course led to
deadlocks. However, there still is the issue of Direct IO reads with
FreeBSD returning EFAULT on occasions because the user pages were not
mapped. This still has to be fully investigated and fixed.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>