Updating zpool iostats
I needed to update how the zpool iostats labels were being done. I did
not do this correctly before. I also got rid of the statistic for
rebuild bulk aggregation writes. We can not do aggregation in the queues
because we are using a seperate lock for the rebuild bulk writes to
allow other IO to go through the queues while rebuild bulk writes are
being issued.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>