Fixing panic issue
When storing rebuild writes we do not place them in the ZIO type AVL
tree for aggregration. So we have to make sure not to try and remove
them from the type tree in vdev_queue_io_remove() in the event the ZIO
was not added in vdev_queue_io_add(). We can check wether the ZIO ABD is
a copy of the parent in this case to avoid removing a ZIO that is not in
the type tree.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>