Fixing race condition
When checking in dmu_buf_undirty() if the dirty record TXT is syncing
the SPA config must be protected under the SCL_CONFIG. This need to be
done in order to make sure we do not transition to sync context for the
TXG from dmu_buf_undirty() to dbuf_undirty(). This was caught with the
ASSERT in dbuf_undirty():

db->db_objset == dmu_objset_pool(db->db_objset)->dp_meta_objset ||
txg != spa_syncing_txg(dmu_objset_spa(db->db_objset))

As the comment before that ASSERT states, dbuf_dirty() can only be
called in open context, unless we are operating on the MOS. Since all
O_DIRECT writes operate in open context, we still much ensure that of a
dirty record is in the process of being sync'ed out, we must allow that
to proceed before removing any asscociated ARC buf and updating the
dbuf's BP. See dmu_write_direct_done() for more information on this.

In order to make this work correctly, the config lock must be held when
we check the condition of whether or not the directory record TXG is
currently syncing out.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>