Fixing issue for failing ZTS test trim_l2arc
PR #12285 introduced a new module parameter (l2arc_exclude_special) to
allow for special or dedup allocation class to specify the L2 ARC should
not be used. However, we pass along the BP for the dbuf from dbuf_read()
to dbuf_read_impl() to account for Direct IO writes. The previous call
for dbuf_is_l2cacheable() directly used the db->db_blkptr so it did not
take into account the BP passed from dbuf_read_impl(). I updated this so
the BP is now passed into this function. If the BP passed is NULL, then
the default behavior of dbuf_is_l2cacheable() remains the same by just
using the db->db_blkptr.

However, the test failure was being caused by trim_l2arc.ksh setting
DIRECT=1 before calling random_reads.fio to fill up the L2 ARC. This
caused Direct IO reads to take place only filling up the L2 ARC with
indirect blocks instead of data blocks. This ultimately led to a failure
for this test due to verify_trim_io getting:
Too few trim IOs issued 2/5
So I update the test case to not use Direct IO as we are wanting to fill
up the L2 ARC with data buffers using random_reads.fio. This allows for
the logic of checking the number of trims to be correct now.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>