Account for Compound Pages in Linux
When attempting to check for if a page was in writeback for Linux we
would deadlock do the fact some pages were compound pages (transparent
huge pages/huge pages). So in the event that we have a compound page, we
only need to check and set the compound_head() page in order to avoid
deadlock. By waiting and adjusting the writeback bit in the compound
page head we are accounting for all the other pages attached to the head
of the compound page.

This deadlock was noticed by doing direct IO writes with recordsizes >
1M, so I added a ZTS test to check for using max recordsizes > 1M for
direct IO.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>