Updating pages to use private field
In order to account for not only compound pages but also merged pages I
started using the pages private field when doing Direct IO writes. This
allows us to isolate when we have already encountered a page, whether it
be a compound page or merged, and only mark the page for writeback one
time.

The way to see the pages have been merged was done by putting a known
data pattern in the user buffer and replicating till the entire buffer
was filled. This resulted in pages being merged (pointing at a single
page that contains the same contents). Using the pages private field
allows to handle this.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>