Adding ZERO_PAGE detection
A user page can come in as  ZERO_PAGE for Direct IO. It is important to
detect this with writes. We do not have to make ZERO_PAGE stable for
writes. Instead we can simply just allocate a zero'ed out page
ourselves. The allocation is important, because if the user changes the
contents of the page it will no longer point at ZERO_PAGE.

One consequence of this is that certain architectures (in particlar
MIPS) has empty_zero_page, which is what ZERO_PAGE references, as a GPL
export. In this case, we can not dectect when a page points to
ZERO_PAGE(0) and We don't allow for either Direct IO writes and reads
in this case.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>