Adjusting ZTS direct tests setup script
With the addition of 56fa4aa compression is now on by default. This
however led to a problem where a new variant of default_raidz_setup
needed to be added for not exiting 1b609d4. Since we test compression
independantly with Direct IO, we need to set compression=off in the
setup script. This means we have to use default_raidz_setup_noexit.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>