Adding module parameter to verify DIO checksum
The added module parameter:
zio_direct_write_verify
Allows for users to opt into a checksum verify to take place after a
Direct IO write has been issued but before it is committed to a VDEV (BP
is updated to point at Direct IO write block). This allows us to avoid
checksum errors in the event the user is manipulating the contents of
their buffer while the Direct IO write is going the the ZIO pipeline.

In the event the user is manipulating the buffer and
zio_direct_write_verify is enabled (set to 1), then a zevent is recorded
as ereport.fs.zfs.dio_verify. Also the top level VDEV verify error
counts can be seen with:
zpool status -d
EINVAL will also be returned to the user in this case.

The other module parameter that I added was:
zfs_dio_write_verify_events_per_second
which is set to 20 by default. This is used like the other rate limiting
events so ZED is not flooded with events if the user is constantly
manipulating their buffer.

I updated the man pages for zpool-events.8 and zfs.4 to explain each
additional module parameter and zevent.

With the addition of this checksum verify I also accounted for what
should happen in the event of a Direct IO write failure. This was never
accounted for in the past. There are two cases that needed to be
handled.

1. The failure happen and there is no ARC buffer associated with this
write. In this case we just undirty the dbuf and set it to DB_UNCACHED.
2. The failure happen and there is an ARC buffer assocated with the
write from a previous buffered write. In this case we must undirty the
dbuf to remove anything from the failed Direct IO write and then redirty
it again. This way all future reads will still grab the valid ARC buffer
that was is assocatied with the dbuf.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>