Added ZTS for direct write checksum verify
This is still a continuation of the WIP from the previous commit. In
this part I added a ZTS test case to run manipulate_user_buffer. At
present this code will set FreeBSD Zpools in SUSPENDED state so that
needs to be investigated more. Also for some strange reason I had to
turn compiler optimizations off in order to get the pthread code for
manipulating the user to buffer to fall out of the while loop. This also
needs further investigation as well.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>