Removing unnecessary code and export symbols
Previously we were exporting dmu_assign_arcbuf_by_dnode() for Lustre,
but this never really made sense to allocate an ARC buf just to do a
Direct IO write. The cost of the allocation and deallocation of the ARC
buf is not necessary. In reality the Lustre code just needs to be
updated to properly do a Direct IO write.

In the vein of updating Lustre to correctly do a Direct IO write I also
exported dmu_write_uio_direct(). The idea is we will probably use a
zfs_uio_t struct in Lustre. I also exported dmu_read_uio_direct() so
that can also be used by Lustre when using a zfs_uio_t struct.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>