Fixing issue that was causing kernel panic
So my code in the cleanup in the event of an IO error in
dmu_write_direct_done() was not right. I originally was calling
dbuf_unoverride() in dbuf_undirty(). However this is not correct. The
reason for this is dbuf_unoverride() would call zio_free() and put the
dirty record BP on the SPA's spa_free_bplist. This would eventually
cause the panic below.

The stack trace was:
txg_sync_thread() -> spa_sync() -> spa_sync_iterate_to convergence() ->
spa_sync_frees() -> zio_free_sync() -> metaslab_check_free() ->
metaslab_check_free_impl() -> range_tree_verify_not_present()

Kernel panic - not syncing segment already in tree; rs=#

The dr_overridden_by was never set when we encountered an IO error with
a Direct IO write. This is by design. So by allowing dbuf_unoverride()
to be called I was inserting the metaslab free for a BP that was not
ever set. Also, the code in zio_done() will cleanup the metaslabs in the
event of an error.

I also cleaned up the dio_write_verify.ksh ZTS test case. I now make
sure to clear out the zevents after the test case so the account is
right for checking the number of Direct IO write verify failures in the
zevents.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>