Fixing race condition again
Originally I thought I had addressed the race condition in dbuf_undirty
when cleaning up after a Direct IO write that also has a ARC buffer
associated with it in b107a54. However, there still is a race condition
and the ASSERT() in dbuf_undirty:

db->db_objset == dmu_objset_pool(db->db_objset)->db_meta_objset || txg
!= spa_syncing_txg(dmu_objset_psa(db->db_objset))

In order to resolve this the SPA config lock must be held the entire
time if we are undirtying a dirty record in dmu_buf_undirty. This is
neccessary so the TXG associated with the dirty record does not move to
the sync phase before the record is undirtied.

Holding the SPA config lock throughout the entire dmu_buf_undirty call
is fine as this should be a rare occurrence that only happens in the
event of mixing buffered and direct IO operations or if there was an IO
failure with a direct IO write.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>