Making sure to update the UIO start offset
When doing O_APPEND the start offset needs to be updated correctly in
zfs_write() after grabbing the range lock. This was an oversight that I
did not notice till running the file_append ZTS test in FreeBSD.
Normally in Linux the kiocb->ki_pos is updated correctly in the ZPL
layers so I did not notice this was an issue. However, this should
always be updated after grabbing the range lock and when doing O_APPEND
no matter what.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>