Removing dead code
I removed all code for placing O_DIRECT write pages in PG_writeback. In
theory, this should not be necessary. We are pinning the the user pages
in through the kernel, so there should be no way for them to be faulted
out (again IN THEORY). Also, after these changes, O_DIRECT now works
with a QEMU VM when doing directsync with the virutal disk cache modes.

As part of these changes I also remove the PG_writeback checks from both
mappedread and zfs_fillpage. Since we are no longer placing the pages
into writeback, these additonal checks are no necessary any longer.

As part of this whole cleanup, I also remove all the previous KSM and
compound page code. In reality, I believe I had an error with the
compound page anyways when free'ing the pages. I believe I needed to
still call put_page() for every page even if it was a compound page. I
am not 100% certain about this, but I believe a reference for the
compound_head() for each page when it is pinned by the kernel though
get_user_pages().

Because we no long need to worry about placing pages in writeback, I
also removed the requirement for ZERO_PAGE() to be available in order to
do O_DIRECT. While I still have code there to check for it if it is
available, this is more just to prevent the user from changing the
contents of before issuing out the write. This at least is a small stop
gap for the zero page, but we can no prevent anything else. That still
would have to be controlled through the zio_direct_write_verify module
parameter.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>