Fixing minor errors man page and _LP64
I had some errors that were caught by doing mandoc.

Also, there was an error when building in the CI for i386 and arm
because I did not make sure to add in a check for _LP64 when trying to
mark a page that we allocate in case it was pointing to ZERO_PAGE(0).
The error was the following without this update:
error: large integer implicitly truncated to unsigned type
[-Werror=overflow]
define set_page_private(page, v) ((page)->private = (v))
/var/lib/buildbot/slaves/buildslave2/Ubuntu_18_04_i386__BUILD_/build
/zfs/module/os/linux/zfs/zfs_uio.c:395:2: note: in expansion of macro
‘set_page_private’
set_page_private(page, ZFS_MARKED_PAGE);

To remedy this I just now make sure we only check for the ZERO_PAGE(0)
if the event we _LP64 is also defined. Otherwise, we will no longer
do this check. Again, this is not a big deal as it is just a way to
prevent users from changing the contents of a page that is all zero's
while it is being written out. This is a rare thing anyways, so it
really is just an optimization that is not necessary for things to
work correct.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>