Adding printf and test to find FreeBSD issue
Currently we are seeing that vm_fault_quick_hold_pages() returns -1
when doing O_DIRECT reads occasionally. I have added printf statments
to trace the failure down. We are the ones setting EFAULT as we are
not getting the expected number of pages when we go to fault and
hold them.

I added a new ZTS test case that can be run by doing:
./scripts/zfs-tests.sh -T direct_freebsd -I 1 -vx

That test will purposely fail just to make sure dmesg output is
put into the stdout file for the test. The dmesg output shows the
printf statements to help with identifying the issue.

From what I can tell, the first direct IO read fails in
vm_fault_quick_hold_pages(). Afterward, we return EFAULT. This causes
FreeBSD to break up the O_DIRECT request into 2 smaller 64K requests
that are successful.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>