Adding ZTS test for FreeBSD stable pages
I fixed the issue of always issuing a buffered write even when no error
occurred with a O_DIRECT write with 3b51977. After doing this stable
pages worked correctly for FreeBSD O_DIRECT writes. I just did not add
a test to verify this. With this commit I have added a ZTS test for
FreeBSD only to test that stable pages work for FreeBSD.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>