Fixing error with mmaped file and O_DIRECT writes
We can remove the UIO_DIRECT flag temporarily while the block size is
growing. This can lead to an issue in zfs_write() causing update_pages()
to be called if the range has become mmap'ed. In order to resolve this
issue, we make sure not to call update_pages() if we are just
temporarily deferring on the O_DIRECT write.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>