Applying OpenZFS Leadership Meeting feedback
Adding in suggestions from the OpenZFS Leadership meeting from June
2022.

The first suggestion was that instead of using a boolean module
parameter for determining when to perform a checksum verify with Direct
I/O writes, we instead switch over to a count based module parameter.
This at least allows for checksum verifies to take place every so often
without incuring the penalty for every write. I have updated the module
parameter to handle this to be zfs_vdev_direct_wite_verify_cnt. By
default, this is set to 100. This also only applies to Linux and at the
top-level VDEV. This is necessary because in Linux we can not place user
pages into write protection, because those pages can come in as
anonymous pages (AKA they are not backed by a file). For FreeBSD, this
is not necessary because user pages can be placed in write protection.
This module parameter determines the number of Direct I/O writes to a
top-level VDEV before a checksum verify should occur and before
commiting the write to the BP. If there is a checksum verify error then
EINVAL is returned. This can somewhat help correlate checksum failures
from the VDEVs back to Direct I/O write checksum verify errors.
These errors can be seen by doing zpool status -d.

As part of this, we also need to make sure that when an ABD is flagged
with ABD_FLAG_FROM_PAGES we actualy copy the contents of the pages when
calling abd_borrow_buf_copy(). This also requires abd_borrow_buf() to
always allocate a buffer to copy the pages into. Finally we need to also
take this into account for abd_return_buf_copy() and abd_return_buf().
This is only relevant to Linux as we can not guarantee the user has not
manipulated the contents of the pages. In order to handle this,
abd_buffor_buf_copy(), abd_borrow_buf(), abd_return_buf_copy(), and
abd_return_buf() were moved out of the general ABD code into OS specfic
ABD code. Normaly ABD's flagged with ABD_FLAG_FROM_PAGES always copied
their data because they are scatter ABDs, but we do allow for
LINEAR_PAGE ABDs when the size of ABD is less than a single page.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>