Hardening Direct I/O ZTS tests
I realized the FIO command in the dio_and_verify function was not quite
doing what we thought it was. I updated it so the verification phase is
done correctly. I also make sure that when we call dio_and_verify we
always run rw, randrw, and write Direct I/O workloads.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>