Updating call to zn_has_cached_data()
Commit #320f0c6 introduced fadvise for Linux. In the new code path,
zn_has_cached_data() is used to check if generic_fadvise() should be
used. Because we expanded to call filemap_range_has_page() if available,
the call to zn_has_cached_data() in zpl_fadvise() needed to be updated
to account for the starting and ending offsets.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>