Updating stable and checksum verify ZTS tests
There was a bizarre error that happen on CentOS Stream 9. The error
occurred when doing a read back after only have Direct I/O checksum
verifies happen every 10th write. The data was read back, but there was
no checksum errors. Not exactly sure why this was the case. However, I
decided to harden these test cases. In order to do this I did the
followoing:
1. Use the default 100 for the module parameter
   zfs_vdev_direct_write_verify_cnt (100). This way we should have more
   checksum verify failures on Linux.
2. Use dd to read back the file. This is important because if the file
   is read using prefetching, then CKSUM errors are not reported to
   zpool status.
3. I make sure we are now doing multiple iterations of the tests. this
   will allow us to identify if there is an underlying issue more
   quickly if we continue to see test case failures.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>