DEBUGGING
Trying to figure out why setting zfs_vdev_direct_write_verify_cnt to
anything besides 1 results in EINVAL never being returned.

Finally figured this out. So we have to make sure to allow the error to
propagate from the VDEV child ZIO in `zio_dio_checksum_verify()` to the
logical parent ZIO. In order to accomplish this, the flag
ZIO_FLAG_DONT_PROPAGATE mush be removed from the VDEV child. This allows
EINVAL to get set in the parent in `zio_notify_parent()`.

Also as part of this, we have to make sure the logical parent no longer
tries to reissue the I/O via `zio_vdev_io_assess()`. The way to do this
is by adding the flag ZIO_FLAG_DONT_RETRY to the logical parent ZIO.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>