Fixing issue with zpool suspend
We need to make sure to pass ZIO_FLAG_CANFAIL to zio_read() in
dmu_read_abd(). The reason for this is we do not want to fall into
zio_reexecute() in zio_done() and get to a place where the zpool is
suspsended. This was observed when using manipulate_user_buffer and the
doing:
dd if=/path_to_file of=/dev/null bs=128k count=# iflag=direct

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>